Added Pytest Plugin to merge two image generation commands into one
