+++ lib/matplotlib/tests/conftest.py
@@ -2,5 +2,4 @@ from matplotlib.testing.conftest import (mpl_test_settings,
                                          mpl_image_comparison_parameters,
                                          pytest_configure,
                                          pytest_unconfigure,
-                                         pd,
-                                         pytest_collection_modifyitems)
+                                         pd)
