+++ lib/mpl_toolkits/tests/conftest.py
@@ -1,6 +1,4 @@
 from matplotlib.testing.conftest import (mpl_test_settings,
                                          mpl_image_comparison_parameters,
                                          pytest_configure,
-                                         pytest_unconfigure,
-                                         pytest_collection_modifyitems,
-                                         pytest_addoption)
+                                         pytest_unconfigure)
