+++ setup.py
@@ -246,6 +246,8 @@ setup(  # Finally, pass this all along to distutils to do the heavy lifting.
     author_email="matplotlib-users@python.org",
     url="https://matplotlib.org",
     download_url="https://matplotlib.org/users/installing.html",
+    entry_points={"pytest11": ["generate_images_plugin = "
+                               "matplotlib.testing.generate_images_plugin"]},
     project_urls={
         'Documentation': 'https://matplotlib.org',
         'Source Code': 'https://github.com/matplotlib/matplotlib',
